// frontend/src/api/models/ScheduledTaskExecutionOut.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ScheduledTaskExecutionOut = {
    id: number;
    task_id: number;
    status: string;
    result: (string | null);
    email_sent: boolean;
    error: (string | null);
    started_at: string;
    completed_at: (string | null);
    created_at: string;
};